server: autoscale JVM heap options
rationale: the jvm settings for graylog server can be automatically
scaled based on the available memory. We set min & max heap to memory/2,
capping it at below 32G, to ensure the JVM will be able to use
-XX:+UseCompressedOops (see
https://www.elastic.co/guide/en/elasticsearch/guide/current/heap-sizing.html)
By default, PermSize and MaxPermSize is also set to the same value.
We leave the GC settings alone, since they are very sensible values for
Graylog2. However, we allow to provide additional java options, such as
-javaagent, or jmx options, which are useful for monitoring.

While we try hard to preserve backwards compatibility in the defaults
files, but this will cause restarts in *default* installations.